pyproject.toml
[tool.poetry]
name = "PSY-TaLiRo"
version = "1.0.0"
description = "System-level verification library using STL"
authors = ["Quinn Thibeault <qthibeau@asu.edu>", "Jacob Anderson <andersonjwan@asu.edu>"]
license = "BSD-3-Clause"
packages = [
    {include = "staliro", from = "src"}
]

[tool.poetry.dependencies]
python = ">=3.7.1, <3.10"
antlr4-python3-runtime = "~=4.5.3"
attrs = "^20.3.0"
numpy = "^1.20.2"
scipy = "^1.6.2"
tltk-mtl = "^0.0.27"
pydantic = "^1.8.2"
typing-extensions = {version = "^3.7.4", python = "<3.8"}
rtamt = {git = "https://github.com/andersonjwan/rtamt", branch="master"}
Sphinx = {version = "^3.5.3", optional = true}
sphinx-autodocgen = {version = "^1.2", optional = true}
pandas = {version = "^1.2.3", optional = true}

[tool.poetry.dev-dependencies]
mypy = "^0.812"
flake8 = "^3.9.0"
flake8-comprehensions = "^3.4.0"
flake8-pyi = "^20.10.0"
flake8-return = "^1.1.2"
flake8-quotes = "^3.2.0"
flake8-variables-names = "^0.0.4"
pep8-naming = "^0.11.1"
flake8-use-fstring = "^1.1"
flake8-import-order = "^0.18.1"
flake8-annotations = "^2.6.2"
flake8-annotations-complexity = "^0.0.6"
flake8-todo = "^0.7"
black = "^21.5b1"

[tool.poetry.extras]
docs = ["sphinx", "sphinx-autodocgen"]
tests = ["pandas"]

[tool.black]
line-length = 99

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
